requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
plotly>=5.18
snowflake-connector-python>=3.6
python-dotenv>=1.0
pyyaml>=6.0
pytest>=8.0
